изменен тип идентификатора на long

+++ sandbox.controller/src/kz/greetgo/sandbox/controller/model/Client.java
@@ -3,9 +3,9 @@ package kz.greetgo.sandbox.controller.model;
 import java.util.Date;
 
 public class Client {
-    public int id;
+    public Long id;
     public String surname, name, patronymic;
     public Gender gender;
     public Date birth_date;
-    public int charm;
+    public Long charm;
 }
