+++ sandbox.controller/src/kz/greetgo/sandbox/controller/model/ClientAccount.java
@@ -3,8 +3,8 @@ package kz.greetgo.sandbox.controller.model;
 import java.sql.Timestamp;
 
 public class ClientAccount {
-    public int id;
-    public int client;
+    public Long id;
+    public Long client;
     public float money;
     public String number;
     public Timestamp registered_at;
