+++ sandbox.controller/src/kz/greetgo/sandbox/controller/model/ClientAccountTransaction.java
@@ -3,9 +3,9 @@ package kz.greetgo.sandbox.controller.model;
 import java.sql.Timestamp;
 
 public class ClientAccountTransaction {
-    public int id;
-    public int account;
+    public Long id;
+    public Long account;
     public float money;
     public Timestamp finished_at;
-    public int type;
+    public Long type;
 }
